add `defaultConfigPath` to managerBasic opts
additionally
* rollback changes in e2e since be can skip loading a
  config
* create `getExpectedConfig` helper in `test-e2e/utils.js`
